Fixed QA issues
