+++ src/renderer/lib/_utils.ts
@@ -6,46 +6,6 @@ export namespace Grid {
     return value[0].toUpperCase() + value.slice(1, value.length);
   }
 
-  export function isParenthesisClosed(value: string) {
-    const pairs = [
-      { start: "(", end: ")" },
-      { start: "[", end: "]" },
-      { start: "{", end: "}" },
-    ];
-    const stacks = new Map();
-
-    // Initialize stacks for each pair
-    pairs.forEach((pair) => {
-      stacks.set(pair, []);
-    });
-
-    // Process each character in the input value
-    for (const char of value) {
-      // Find the corresponding pair for the current character
-      const pair = pairs.find((e) => e.start === char || e.end === char);
-
-      // If no pair is found (invalid character), continue
-      if (!pair) continue;
-
-      // Check if the character is a start or end bracket for the pair
-      switch (char) {
-        case pair.start:
-          stacks.get(pair).push(char); // Push to the stack of the corresponding pair
-          break;
-        case pair.end:
-          // Check if there's a corresponding start, and pop from the stack
-          if (stacks.get(pair).length === 0) {
-            return false; // Unmatched closing bracket
-          }
-          stacks.get(pair).pop();
-          break;
-      }
-    }
-
-    // Check if all stacks are empty, meaning all parentheses are closed correctly
-    return [...stacks.values()].every((stack) => stack.length === 0);
-  }
-
   export function getClosestEvent(events: number[], event: number) {
     if (events.map((e) => Number(e)).includes(Number(event))) {
       return event;
