+++ src/renderer/runtime/runtime.ts
@@ -351,15 +351,6 @@ export class GridAction extends RuntimeNode<ActionData> {
     const parent = this.parent as GridEvent;
     const diff = data.toLua().length - this.data.toLua().length;
 
-    if (!Grid.isParenthesisClosed(data.script)) {
-      return Promise.reject({
-        value: false,
-        text: Runtime.ErrorText.UNCLOSED_PARENTHESIS,
-        type: GridOperationType.UPDATE_ACTION,
-        info: (this.parent as GridEvent)?.getInfo(),
-      });
-    }
-
     if (parent.getAvailableChars() - diff >= 0) {
       this.script = data.script;
       this.short = data.short;
